Add feature to publish transform from gnss origin to map origin
The old behavior of not publishing the transformation (hence, exiting
the program once the conversion is done) is preserved unless both
gnss_frame_id and map_frame_id are specified. If both these parameters
are specified (like in the csv_to_dem.yaml file), the node will not exit
after the conversion is done.